fix for dvquantity.magnitude not being parsed correctly after checkout #31

+++ grails-app/domain/registros/valores/DvQuantity.groovy
@@ -2,6 +2,6 @@ package registros.valores
 
 class DvQuantity extends DataValue {
 
-   float magnitude
+   BigDecimal magnitude // was float but I had some issues: http://stackoverflow.com/questions/37450774/grails-2-5-3-not-binding-float-fields-correctly/37454016
    String units
 }
